blog/requirements.txt
crispy-tailwind==1.0.3
dj-stripe==2.8.3
Django==5.0.2
django-crispy-forms==2.1
Jinja2==3.1.3
pillow==10.2.0
qrcode==7.4.2
regex==2023.12.25
stripe==8.3.0
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0